Updated code and added authorization for all repositories. Updated permissions file

+++ Library.Services/BookService.cs
@@ -1,18 +1,13 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Library.BusinessObjects;
+﻿using Library.BusinessObjects;
+using Library.core;
 using Library.Repository;
-
-
 
 namespace Library.Services
 {
     public class BookService : IBookService
     {
-        
+        private AuthorizationService authorizationService = new AuthorizationService();
+        private User loggedInUser = System.Threading.Thread.CurrentPrincipal.GetLoggedInUser();
         private IBookRepository _bookRepository;
 
         public BookService(IBookRepository bookRepository)
@@ -20,34 +15,44 @@ namespace Library.Services
         {
             _bookRepository = bookRepository;
         }
-        
+
         //BookRepository bookRepository = new BookRepository();
         public bool AddBook(Book book)
         {
-            return _bookRepository.AddBook(book);
+            if (authorizationService.Authorize(loggedInUser.roleName, "AddBook"))
+                return _bookRepository.AddBook(book);
+            else
+                return false;
         }
-        public Book ManageBooks() {
-            return new Book();
-        }
+
         public Book SearchBookByName(string name)
         {
-         return _bookRepository.SearchBookByName(name);
+            if (authorizationService.Authorize(loggedInUser.roleName, "SearchBookByName"))
+                return _bookRepository.SearchBookByName(name);
+            else
+                return null;
         }
 
-        public Book SearchBookByPublishedBy(String PublishedBy)
+        public Book SearchBookByPublishedBy(string publishedBy)
         {
-            return _bookRepository.SearchByPublishedBy(PublishedBy);
+            if (authorizationService.Authorize(loggedInUser.roleName, "SearchBookByPublishedBy"))
+
+                return _bookRepository.SearchByPublishedBy(publishedBy);
+            else
+                return null;
         }
 
-        public bool EditQuantity(int bookID, int Quantity)
+        public bool EditQuantity(int bookID, int quantity)
         {
-            return _bookRepository.EditQuantity(bookID, Quantity);
+            return _bookRepository.EditQuantity(bookID, quantity);
         }
-
 
         public bool DeleteBook(int bookID)
         {
-            return _bookRepository.DeleteBook(bookID);
+            if (authorizationService.Authorize(loggedInUser.roleName, "DeleteBook"))
+                return _bookRepository.DeleteBook(bookID);
+            else
+                return false;
         }
     }
 }
